Code changes corresponding to template move
